requirements.txt
python-telegram-bot
torrentool
google-auth-oauthlib
google-api-python-client
python-dotenv
requests
